+++ config/custom-logo.php
@@ -1,0 +1,18 @@
+<?php
+/**
+ * Genesis Sample child theme.
+ *
+ * @author  StudioPress
+ * @license GPL-2.0-or-later
+ * @link    https://my.studiopress.com/themes/genesis-sample/
+ */
+
+/**
+ * Custom Logo configuration.
+ */
+return array(
+	'height'      => 120,
+	'width'       => 700,
+	'flex-height' => true,
+	'flex-width'  => true,
+);