+++ config/editor-color-palette.php
@@ -1,0 +1,29 @@
+<?php
+/**
+ * Genesis Sample child theme.
+ *
+ * @author  StudioPress
+ * @license GPL-2.0-or-later
+ * @link    https://my.studiopress.com/themes/genesis-sample/
+ */
+
+/**
+ * Editor color palette config.
+ */
+return array(
+	array(
+		'name'  => __( 'Light gray', 'genesis-sample' ),
+		'slug'  => 'light-gray',
+		'color' => '#f5f5f5',
+	),
+	array(
+		'name'  => __( 'Medium gray', 'genesis-sample' ),
+		'slug'  => 'medium-gray',
+		'color' => '#999',
+	),
+	array(
+		'name'  => __( 'Dark gray', 'genesis-sample' ),
+		'slug'  => 'dark-gray',
+		'color' => '#333',
+	),
+);