+++ config/editor-font-sizes.php
@@ -1,0 +1,34 @@
+<?php
+/**
+ * Genesis Sample child theme.
+ *
+ * @author  StudioPress
+ * @license GPL-2.0-or-later
+ * @link    https://my.studiopress.com/themes/genesis-sample/
+ */
+
+/**
+ * Editor font sizes config.
+ */
+return array(
+	array(
+		'name'      => __( 'Small', 'genesis-sample' ),
+		'size'      => 12,
+		'slug'      => 'small',
+	),
+	array(
+		'name'      => __( 'Normal', 'genesis-sample' ),
+		'size'      => 16,
+		'slug'      => 'normal',
+	),
+	array(
+		'name'      => __( 'Large', 'genesis-sample' ),
+		'size'      => 20,
+		'slug'      => 'large',
+	),
+	array(
+		'name'      => __( 'Larger', 'genesis-sample' ),
+		'size'      => 24,
+		'slug'      => 'larger',
+	),
+);