+++ config/menus.php
@@ -1,0 +1,16 @@
+<?php
+/**
+ * Genesis Sample child theme.
+ *
+ * @author  StudioPress
+ * @license GPL-2.0-or-later
+ * @link    https://my.studiopress.com/themes/genesis-sample/
+ */
+
+/**
+ * Supported Genesis navigation menus.
+ */
+return array(
+	'primary'   => __( 'Header Menu', 'genesis-sample' ),
+	'secondary' => __( 'Footer Menu', 'genesis-sample' ),
+);