+++ template-blocks.php
@@ -1,0 +1,27 @@
+<?php
+/**
+ * Genesis Sample.
+ *
+ * A template to force full-width layout, remove breadcrumbs, and remove the page title.
+ *
+ * Template Name: Blocks
+ *
+ * @package Genesis Sample
+ * @author  StudioPress
+ * @license GPL-2.0-or-later
+ * @link    https://www.studiopress.com/
+ */
+
+// Removes the entry header markup and page title.
+remove_action( 'genesis_entry_header', 'genesis_entry_header_markup_open', 5 );
+remove_action( 'genesis_entry_header', 'genesis_entry_header_markup_close', 15 );
+remove_action( 'genesis_entry_header', 'genesis_do_post_title' );
+
+// Forces full width content layout.
+add_filter( 'genesis_site_layout', '__genesis_return_full_width_content' );
+
+// Removes the breadcrumbs.
+remove_action( 'genesis_before_loop', 'genesis_do_breadcrumbs' );
+
+// Runs the Genesis loop.
+genesis();